Add MapOption and UnmapOption types
Define interfaces for controlling Go-to-Tony and Tony-to-Go mapping behavior,
separate from encoding/parsing options. Config structs include fields for
passing through EncodeOptions and ParseOptions to their respective stages.

WIP: Update to.go signatures to use MapOption

Changed ToTony, ToTonyIR, and all reflection helpers to use MapOption
instead of encode.EncodeOption. Still need to update mapper_to.go and
from.go in next commit.

Fix critical bugs #2 and #7

Bug #2: Fix incorrect method name ir.FromFloat64 → ir.FromFloat in codegen
Bug #7: Fix sparse array type to use map[uint32]T instead of map[int]T

Fix Bug #3: LineComment capitalization inconsistency

Changed 'LineComment' to 'lineComment' for consistency with other tag names

Fix codegen to skip embedded schema marker fields and update tests

- Skip embedded fields with schema= or schemagen= tags during codegen
- These are marker fields for struct-level configuration, not actual data fields
- Update test expectations for Bug #7 (sparse array uint32) and Bug #3 (lineComment lowercase)
- All gomap tests now pass

Update codegen to generate gomap.MapOption/UnmapOption signatures

- Changed generated method signatures to use gomap.MapOption and gomap.UnmapOption
- Added gomap import to generated files
- Updated ToTony to use gomap.ToEncodeOptions(opts...)
- Updated FromTony to use gomap.ToParseOptions(opts...)
- Note: Bug #8 (isZeroValue/assign logic) still exists and causes compilation errors

Fix Bug #8: Properly handle node/err variable declarations in codegen

- Declare var node and err at function start instead of using :=
- Always use = for assignments since vars are pre-declared
- Conditionally declare vars only if struct has fields that need them
- Exclude *ir.Node fields from needsVars check (they don't need conversion)
- This fixes 'undefined: node/err' and 'declared and not used' errors
- All dirbuild tests now pass

Update codegen tests to expect new MapOption/UnmapOption signatures

- Updated all test assertions to expect gomap.MapOption and gomap.UnmapOption
- Updated tests to expect gomap.ToEncodeOptions() and gomap.ToParseOptions() calls
- All gomap and gomap/codegen tests now pass

Fix unused variable declaration in codegen for map[string]*ir.Node

- Refined needsVars check in generator.go to be more specific
- Now checks element types for Slices and Maps
- Explicitly excludes *ir.Node values in Maps from triggering var declaration
- This fixes 'declared and not used' errors for structs like Body in api package

resync

marshal text and misc issues

checkpoint

+++ go-tony/gomap/type_extract.go
@@ -94,7 +94,7 @@ func isNullableFromOr(def *ir.Node) bool {
 //   - "!or [null, string]" → returns reflect.PtrTo(reflect.TypeOf(""))
 //   - ".array(string)" → returns reflect.SliceOf(reflect.TypeOf("")) ([]string)
 //   - ".array(.array(string))" → returns [][]string
-//   - ".sparsearray(string)" → returns map[int]string
+//   - ".sparsearray(string)" → returns map[uint32]string
 //
 // The function recursively resolves references and builds up complex types, including
 // nested parameterized types.
@@ -366,7 +366,7 @@ func extractGoTypeFromOr(def *ir.Node, s *schema.Schema, registry *schema.Schema
 			nonNullNodes = append(nonNullNodes, elem)
 			continue
 		}
-		
+
 		// Check for null (only if not a reference)
 		if elem.Type == ir.NullType {
 			hasNull = true
@@ -418,7 +418,7 @@ func extractGoTypeFromOrUnion(nonNullNodes []*ir.Node, s *schema.Schema, registr
 
 		// Generate a field name from the type
 		fieldName := typeToFieldName(typ)
-		
+
 		// Ensure unique field names
 		baseName := fieldName
 		counter := 1
@@ -485,16 +485,18 @@ func typeToFieldName(typ reflect.Type) string {
 // 3. Skip constraint nodes (like !not null, !not, etc.)
 //
 // Example:
-//   int: !and
-//     - .[number]        # Base type: number → float64
-//     - int: !not null    # Constraint: skip this
-//   Result: float64
+//
+//	int: !and
+//	  - .[number]        # Base type: number → float64
+//	  - int: !not null    # Constraint: skip this
+//	Result: float64
 //
 // Example with parameterized type:
-//   array(t): !and
-//     - .[array]          # Base type: array
-//     - !all.type t       # Parameter constraint: t is the element type
-//   When used as .array(string), we extract []string
+//
+//	array(t): !and
+//	  - .[array]          # Base type: array
+//	  - !all.type t       # Parameter constraint: t is the element type
+//	When used as .array(string), we extract []string
 func extractGoTypeFromAnd(def *ir.Node, s *schema.Schema, registry *schema.SchemaRegistry) (reflect.Type, error) {
 	if def.Type != ir.ObjectType && def.Type != ir.ArrayType {
 		return nil, fmt.Errorf("!and must be an object or array")
@@ -542,7 +544,7 @@ func extractGoTypeFromAnd(def *ir.Node, s *schema.Schema, registry *schema.Schem
 func extractGoTypeFromTypeTag(def *ir.Node) (reflect.Type, error) {
 	// !irtype tags typically have a value that indicates the type
 	// Examples: !irtype "", !irtype 1, !irtype true, !irtype []
-	
+
 	if def.Type == ir.StringType {
 		// !irtype "" → string
 		return reflect.TypeOf(""), nil
@@ -621,12 +623,12 @@ func extractGoTypeFromParameterizedRef(tag string, s *schema.Schema, registry *s
 		return reflect.SliceOf(elemType), nil
 
 	case "sparsearray":
-		// .sparsearray(T) → map[int]T (sparse arrays are maps with numeric keys)
+		// .sparsearray(T) → map[uint32]T (sparse arrays are maps with uint32 keys)
 		elemType, err := extractElementTypeFromArg(args[0], s, registry)
 		if err != nil {
 			return nil, fmt.Errorf("failed to extract element type from .sparsearray(%s): %w", args[0], err)
 		}
-		return reflect.MapOf(reflect.TypeOf(0), elemType), nil
+		return reflect.MapOf(reflect.TypeOf(uint32(0)), elemType), nil
 
 	default:
 		// Unknown constructor - try to resolve as a definition name
